requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
beautifulsoup4==4.8.2
boto3==1.12.24
botocore==1.15.24
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-email-url==1.0.0
Django==3.0.4
django-cache-url==3.1.2
django-cors-headers==3.2.1
django-graphql-jwt==0.3.1
django-storages==1.9.1
docutils==0.15.2
environs==7.3.1
graphene==2.1.8
graphene-django==2.8.2
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
marshmallow==3.5.1
Pillow==7.0.0
promise==2.3
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.8
webpreview==1.6.0
